perf(i18n): skip loading animation for cached translations
- Track loaded translation bundles separately from in-flight requests
- Show the language download animation only for uncached languages
